requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.5
python-dotenv==1.0.0
rest_framework_simplejwt==0.0.2
